changed specification of default hosts

+++ lib/linecook/config.rb
@@ -2,18 +2,23 @@ module Linecook
   module Config
     module_function
     
-    def hostnames(path)
-      hostnames = []
-      File.open(path) do |io|
-        io.each_line do |line|
-          next unless line =~ /^\s*Host\s+(\w+)/
-          hostnames << $1 
+    def hosts(path)
+      hosts = []
+      
+      if File.exists?(path)
+        File.open(path) do |io|
+          io.each_line do |line|
+            next unless line =~ /^\s*Host\s+(\w+)/
+            hosts << $1 
+          end
         end
       end
-      hostnames
+      
+      hosts
     end
     
     DEFAULT_SSH_CONFIG_FILE = File.expand_path(ENV['LINECOOK_SSH_CONFIG_FILE'] || 'config/ssh')
-    HOSTNAMES = Hash.new {|hash, path| hash[path] = hostnames(path) }
+    DEFAULT_HOSTS = hosts(DEFAULT_SSH_CONFIG_FILE)
+    DEFAULT_HOST  = ENV['LINECOOK_DEFAULT_HOST'] || DEFAULT_HOSTS.first
   end
 end
